Move some files around for sanity
