+++ conda_forge_tick/auto_tick.py
@@ -20,8 +20,6 @@ from .git_utils import (
     get_repo,
     push_repo,
     is_github_api_limit_reached,
-    ensure_label_exists,
-    label_pr,
 )
 from .path_lengths import cyclic_topological_sort
 from .utils import (
@@ -41,7 +39,6 @@ from typing import (
     Tuple,
     Dict,
     Set,
-    Sized,
     MutableMapping,
 )
 
@@ -50,18 +47,15 @@ logger = logging.getLogger("conda_forge_tick.auto_tick")
 # TODO: move this back to the bot file as soon as the source issue is sorted
 # https://travis-ci.org/regro/00-find-feedstocks/jobs/388387895#L1870
 from .utils import frozen_to_json_friendly
+from .contexts import MigratorsContext
 from .migrators import (
-    MigratorsContext,
     Migrator,
     Version,
     PipMigrator,
     LicenseMigrator,
-    Noarch,
-    NoarchR,
     MigrationYaml,
     MigratorContext,
     Rebuild,
-    BlasRebuild,
     Replacement,
     ArchRebuild,
 )
@@ -144,11 +138,12 @@ def run(
     no_noarch_files = [
         f"{script_name}.{ext}" for script_name in script_names for ext in exts
     ]
-    if isinstance(migrator, Noarch) and any(
-        x in os.listdir(recipe_dir) for x in no_noarch_files
-    ):
-        eval_xonsh(f"rm -rf {feedstock_dir}")
-        return False, False
+    # TODO: Remove this
+    # if isinstance(migrator, Noarch) and any(
+    #     x in os.listdir(recipe_dir) for x in no_noarch_files
+    # ):
+    #     eval_xonsh(f"rm -rf {feedstock_dir}")
+    #     return False, False
     # migrate the `meta.yaml`
     migrate_return = migrator.migrate(recipe_dir, feedstock_ctx.attrs, **kwargs)
     if not migrate_return:
@@ -431,58 +426,58 @@ def add_rebuild_successors(
     )
 
 
-def add_rebuild_blas(migrators, gx):
-    """Adds rebuild blas 2.0 migrators.
-
-    Parameters
-    ----------
-    migrators : list of Migrator
-        The list of migrators to run.
-
-    """
-    total_graph = copy.deepcopy(gx)
-
-    for node, node_attrs in gx.nodes.items():
-        attrs: "AttrsTypedDict" = node_attrs["payload"]
-        meta_yaml = attrs.get("meta_yaml", {}) or {}
-        bh = get_requirements(meta_yaml)
-        pkgs = {
-            "openblas",
-            "openblas-devel",
-            "mkl",
-            "mkl-devel",
-            "blas",
-            "lapack",
-            "clapack",
-        }
-        blas_c = len(pkgs.intersection(bh)) > 0
-
-        rq = _host_run_test_dependencies(meta_yaml)
-
-        for e in list(total_graph.in_edges(node)):
-            if e[0] not in rq:
-                total_graph.remove_edge(*e)
-        if not any([blas_c]):
-            pluck(total_graph, node)
-
-    # post plucking we can have several strange cases, lets remove all selfloops
-    total_graph.remove_edges_from(nx.selfloop_edges(total_graph))
-
-    top_level = {
-        node for node in total_graph if not list(total_graph.predecessors(node))
-    }
-    cycles = list(nx.simple_cycles(total_graph))
-
-    migrators.append(
-        BlasRebuild(
-            graph=total_graph,
-            pr_limit=5,
-            name="blas-2.0",
-            top_level=top_level,
-            cycles=cycles,
-            obj_version=0,
-        ),
-    )
+# def add_rebuild_blas(migrators, gx):
+#     """Adds rebuild blas 2.0 migrators.
+#
+#     Parameters
+#     ----------
+#     migrators : list of Migrator
+#         The list of migrators to run.
+#
+#     """
+#     total_graph = copy.deepcopy(gx)
+#
+#     for node, node_attrs in gx.nodes.items():
+#         attrs: "AttrsTypedDict" = node_attrs["payload"]
+#         meta_yaml = attrs.get("meta_yaml", {}) or {}
+#         bh = get_requirements(meta_yaml)
+#         pkgs = {
+#             "openblas",
+#             "openblas-devel",
+#             "mkl",
+#             "mkl-devel",
+#             "blas",
+#             "lapack",
+#             "clapack",
+#         }
+#         blas_c = len(pkgs.intersection(bh)) > 0
+#
+#         rq = _host_run_test_dependencies(meta_yaml)
+#
+#         for e in list(total_graph.in_edges(node)):
+#             if e[0] not in rq:
+#                 total_graph.remove_edge(*e)
+#         if not any([blas_c]):
+#             pluck(total_graph, node)
+#
+#     # post plucking we can have several strange cases, lets remove all selfloops
+#     total_graph.remove_edges_from(nx.selfloop_edges(total_graph))
+#
+#     top_level = {
+#         node for node in total_graph if not list(total_graph.predecessors(node))
+#     }
+#     cycles = list(nx.simple_cycles(total_graph))
+#
+#     migrators.append(
+#         BlasRebuild(
+#             graph=total_graph,
+#             pr_limit=5,
+#             name="blas-2.0",
+#             top_level=top_level,
+#             cycles=cycles,
+#             obj_version=0,
+#         ),
+#     )
 
 
 def add_replacement_migrator(
