+++ conda_forge_tick/test_status.py
@@ -3,10 +3,8 @@ Similar to test-status, but for reports
 """
 from typing import Any
 
-from conda_forge_tick.utils import setup_logger
 from .migrators import Rebuild
 from .auto_tick import initialize_migrators, migrator_status
-import os
 import yaml
